requirements.txt
Authlib==1.1.0
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
cryptography==38.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Login==0.4.1
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.0.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.1.0
python-dotenv==0.21.0
requests==2.21.0
six==1.16.0
urllib3==1.24.3
Werkzeug==2.0.3
gunicorn==20.1.0